requirements.txt
streamlit
Pillow
tensorflow
pandas
numpy
reportlab
opencv-python
scikit-learn